Reformatting Introduced smart_enabled property
