+++ src/main/java/ca/uhn/fhir/jpa/starter/Application.java
@@ -2,8 +2,8 @@ package ca.uhn.fhir.jpa.starter;
 
 import ca.uhn.fhir.jpa.starter.mdm.MdmConfig;
 import ca.uhn.fhir.jpa.starter.annotations.OnEitherVersion;
+import ca.uhn.fhir.jpa.starter.smart.ScopeBasedAuthorizationInterceptor;
 import ca.uhn.fhir.jpa.starter.smart.WellKnownEndpointController;
-import ca.uhn.fhir.jpa.starter.smart.WellknownEndpointHelper;
 import ca.uhn.fhir.jpa.subscription.channel.config.SubscriptionChannelConfig;
 import ca.uhn.fhir.jpa.subscription.match.config.SubscriptionProcessorConfig;
 import ca.uhn.fhir.jpa.subscription.match.config.WebsocketDispatcherConfig;
@@ -26,7 +26,7 @@ import org.springframework.web.servlet.DispatcherServlet;
 @ServletComponentScan(basePackageClasses = {
   JpaRestfulServer.class})
 @SpringBootApplication(exclude = {ElasticsearchRestClientAutoConfiguration.class})
-@Import({SubscriptionSubmitterConfig.class, SubscriptionProcessorConfig.class, SubscriptionChannelConfig.class, WebsocketDispatcherConfig.class, MdmConfig.class})
+@Import({ SubscriptionSubmitterConfig.class, SubscriptionProcessorConfig.class, SubscriptionChannelConfig.class, WebsocketDispatcherConfig.class, MdmConfig.class})
 public class Application extends SpringBootServletInitializer {
 
   public static void main(String[] args) {
