+++ src/main/java/ca/uhn/fhir/jpa/starter/config/FhirServerConfigDstu3.java
@@ -1,25 +1,29 @@
-package ca.uhn.fhir.jpa.starter;
+package ca.uhn.fhir.jpa.starter.config;
 
 import ca.uhn.fhir.context.ConfigurationException;
-import ca.uhn.fhir.jpa.config.BaseJavaConfigDstu2;
+import ca.uhn.fhir.jpa.config.BaseJavaConfigDstu3;
 import ca.uhn.fhir.jpa.search.DatabaseBackedPagingProvider;
-import ca.uhn.fhir.jpa.starter.annotations.OnDSTU2Condition;
+import ca.uhn.fhir.jpa.search.lastn.ElasticsearchSvcImpl;
+import ca.uhn.fhir.jpa.starter.AppProperties;
+import ca.uhn.fhir.jpa.starter.annotations.OnDSTU3Condition;
+import ca.uhn.fhir.jpa.starter.cql.StarterCqlDstu3Config;
+import javax.annotation.PostConstruct;
+import javax.persistence.EntityManagerFactory;
+import javax.sql.DataSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Primary;
 import org.springframework.core.env.ConfigurableEnvironment;
 import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 
-import javax.annotation.PostConstruct;
-import javax.persistence.EntityManagerFactory;
-import javax.sql.DataSource;
-
 @Configuration
-@Conditional(OnDSTU2Condition.class)
-public class FhirServerConfigDstu2 extends BaseJavaConfigDstu2 {
+@Conditional(OnDSTU3Condition.class)
+@Import(StarterCqlDstu3Config.class)
+public class FhirServerConfigDstu3 extends BaseJavaConfigDstu3 {
 
   @Autowired
   private DataSource myDataSource;
@@ -45,6 +49,7 @@ public class FhirServerConfigDstu2 extends BaseJavaConfigDstu2 {
 	  }
   }
 
+
   @Override
   public DatabaseBackedPagingProvider databaseBackedPagingProvider() {
     DatabaseBackedPagingProvider pagingProvider = super.databaseBackedPagingProvider();
@@ -55,9 +60,9 @@ public class FhirServerConfigDstu2 extends BaseJavaConfigDstu2 {
 
   @Autowired
   private ConfigurableEnvironment configurableEnvironment;
-
+  
   @Override
-  @Bean()
+  @Bean
   public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
     LocalContainerEntityManagerFactoryBean retVal = super.entityManagerFactory();
     retVal.setPersistenceUnitName("HAPI_PU");
@@ -67,6 +72,7 @@ public class FhirServerConfigDstu2 extends BaseJavaConfigDstu2 {
     } catch (Exception e) {
       throw new ConfigurationException("Could not set the data source due to a configuration issue", e);
     }
+
     retVal.setJpaProperties(EnvironmentHelper.getHibernateProperties(configurableEnvironment));
     return retVal;
   }
@@ -79,5 +85,23 @@ public class FhirServerConfigDstu2 extends BaseJavaConfigDstu2 {
     return retVal;
   }
 
+  @Bean()
+  public ElasticsearchSvcImpl elasticsearchSvc() {
+    if (EnvironmentHelper.isElasticsearchEnabled(configurableEnvironment)) {
+      String elasticsearchUrl = EnvironmentHelper.getElasticsearchServerUrl(configurableEnvironment);
+      String elasticsearchHost;
+      if (elasticsearchUrl.startsWith("http")) {
+        elasticsearchHost = elasticsearchUrl.substring(elasticsearchUrl.indexOf("://") + 3, elasticsearchUrl.lastIndexOf(":"));
+      } else {
+        elasticsearchHost = elasticsearchUrl.substring(0, elasticsearchUrl.indexOf(":"));
+      }
+      String elasticsearchUsername = EnvironmentHelper.getElasticsearchServerUsername(configurableEnvironment);
+      String elasticsearchPassword = EnvironmentHelper.getElasticsearchServerPassword(configurableEnvironment);
+      int elasticsearchPort = Integer.parseInt(elasticsearchUrl.substring(elasticsearchUrl.lastIndexOf(":")+1));
+      return new ElasticsearchSvcImpl(elasticsearchHost, elasticsearchPort, elasticsearchUsername, elasticsearchPassword);
+    } else {
+      return null;
+    }
+  }
 
 }
