+++ src/main/java/ca/uhn/fhir/jpa/starter/validation/IRepositoryValidationInterceptorFactory.java
@@ -1,4 +1,4 @@
-package ca.uhn.fhir.jpa.starter;
+package ca.uhn.fhir.jpa.starter.validation;
 
 import ca.uhn.fhir.jpa.interceptor.validation.RepositoryValidatingInterceptor;
