+++ src/main/java/ca/uhn/fhir/jpa/starter/validation/RepositoryValidationInterceptorFactoryDstu3.java
@@ -1,4 +1,4 @@
-package ca.uhn.fhir.jpa.starter;
+package ca.uhn.fhir.jpa.starter.validation;
 
 import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.jpa.api.dao.DaoRegistry;
